Initial commit
This bundle provides Twig adaptations of Zurb's Ink email templates. Currently based on version 1.0.2 of the Ink templates.

This commit contains code for initializing the bundle and templates based on the Ink boilerplate & basic templates. Composer configuration is also included.

This bundle is provided under the terms of the MIT license.